backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.2.1
google-cloud-firestore>=2.16.0
firebase-admin>=6.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
mpy-cross>=1.22.0
pybricksdev>=2.3.2
python-dotenv>=1.0.1
